+++ packages/sdk-redux/src/reduxSlices/transactionTrackerSlice/ethersError.ts
@@ -1,7 +1,8 @@
-import {ethers} from 'ethers';
+import {ethers, Transaction} from 'ethers';
 
 export type EthersError = Error & {
     code: ethers.errors;
     reason?: 'reverted' | 'replaced' | 'repriced' | 'cancelled';
     cancelled?: boolean;
+    replacement?: {hash: string} & Transaction;
 };